Reorganize, make a first pass at awesome-list style guide compliance.
This commit makes massive changes to the list organization. Instead of
categorizing the projects as "Tor Project's Projects" or not, mainly,
which was one gigantic and growing list of current and outdated
projects, the line items are categorized by their purpose or function.
The goal is to make the list more useful for someone who is unfamiliar
with Tor and wants to know which tool or project is the most relevant to
their use case.

In addition, this commit adds the following Tor-based tools:

* Ricochet, a Tor Onion service-based Jabber messenger.
* Whonix, a GNU/Linux distribution that torifies everything by default.
* OnionScan, a security tool for alerting on Onion service misconfigs.
* Vanguards, an Onion service entry guard attack mitigation script.
* tor.rb, a Ruby library for interacting with Tor.

This commit further adds a License heading, and edits many list items
for conformance with the awesome-list style guide. Namely, this means
"a" and "an" prepositions have been dropped, project names are removed
from the project line item description, periods are added to the end of
line item descriptions, headings are moved to level one, a
`CONTRIBUTING.md` file has been added, and so on.