Remove Globe; it's not maintained. Remove duplicate names in line items.
The style guide for Awesome Lists requests that line items not repeat
themselves. This commit removes numerous instances where the name of the
line item was repeated as the first part of its description. For
example:

* "TorBirdy - TorBirdy is Torbutton…" -> "TorBirdy - Torbutton…"

This commit also removes Globe, as the Tor project volunteering page
explicitly mentions that it's no longer being maintained. Unmaintained
projects are not to be listed on Awesome Lists, also according to its
style guide.